@@ -1,21 +1,18 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
-import { Provider } from "react-redux";
-import { configureStore } from "@reduxjs/toolkit";
-import rootReducer from "./reducer";
-import { Toaster } from "@/components/ui/toaster"
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
+import { store, persistor } from './store/store';
 
-const store= configureStore({
-  reducer:rootReducer
-})
 
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-      <Provider store={store}>
-        <App />   
-        <Toaster />
-      </Provider>
-  </React.StrictMode>,
-)
+import App from './App';
+import './index.css';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <Provider store={store}>
+    <PersistGate loading={null} persistor={persistor}>
+      <App />
+    </PersistGate>
+  </Provider>
+);
